Move ChessPiece to separate file, create file for future VM
