+++ chess/ChessGameViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  ChessGameViewModel.swift
+//  chess
+//
+//  Created by Yeliena Khaletska on 11.07.2024.
+//
+
+import Foundation
+
+final class ChessGameViewModel {
+
+
+
+}
